move generateArrayOfStrings func to separate file
