+++ js/main.js
@@ -3,35 +3,11 @@ import getRandomFloat from './utils/get-random-float.js';
 import getUniqRandomItemsFromArray from './utils/get-uniq-random-items-from-array.js';
 import makeCounter from './utils/make-counter.js';
 import generateAuthor from './src/generate-author.js';
+import generateArrayOfStrings from './utils/generate-array-of-strings.js';
 
 // offer's helping functions
 
 const getTitleNumber = makeCounter();
-
-/**
- * Function to create Array of concatenated string with counter's value inside.
- *
- * @param {string} stringHead - first part of concatenated string
- * @param {string} stringTail - last part of concatenated string
- * @param {number} rangeStart - starting number to set in the middle of string, included
- * @param {number} rangeEnd - ending number to set in the middle of string, included
- * @param {number} step - increment or decrement(use '-' to decrement value)
- * @return {array} Array of strings with enumerated range inside of each
- */
-
-const generateArrayOfStrings = (stringHead, stringTail, rangeStart, rangeEnd, step = 1) => {
-  const result = [];
-
-  for (
-    let current = rangeStart;
-    step < 0 ? current >= rangeEnd : current <= rangeEnd;
-    current += step
-  ) {
-    result.push(stringHead + current + stringTail);
-  }
-
-  return result;
-}
 
 // offer's data storage
